services/api/requirements.txt
fastapi
uvicorn
pydantic
email-validator
python-dotenv
httpx
orjson
polars
sqlalchemy
psycopg2-binary
slowapi
